requirements.txt
antlr4-python3-runtime==4.9.3
astroid==2.12.13
dill==0.3.6
et-xmlfile==1.1.0
isort==5.11.4
lazy-object-proxy==1.8.0
mccabe==0.7.0
numpy==1.24.1
openpyxl==3.0.10
platformdirs==2.6.2
pylint==2.15.9
scipy==1.9.3
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
wrapt==1.14.1
